Add files via upload
This is a program to create a small quiz game. This game has 10 questions.
All the questions are MCQ type.
Each question will have 1 point
From the 6th to 10th question, there will be a negative marking of 1 point and if the answer is wrong and if the answer is correct 1 point will be added.
You will be out of the game if answers are wrong continuously 3 times.